+++ src/components/ReactComponents/LikeButton.jsx
@@ -1,0 +1,4 @@
+import { customActivityButton } from './ActivityButton';
+
+/** Button to view and perform a "Like" action on an item. */
+export default customActivityButton('Like', 'Like', 'You liked', 'Liked');
